Exception Handling

+++ domainModelORM/src/main/java/com/devsuperior/domainModelORM/services/ProductService.java
@@ -3,10 +3,15 @@ package com.devsuperior.domainModelORM.services;
 import com.devsuperior.domainModelORM.DTO.ProductDTO;
 import com.devsuperior.domainModelORM.entities.Product;
 import com.devsuperior.domainModelORM.repositories.ProductRepository;
+import com.devsuperior.domainModelORM.services.exceptions.DatabaseException;
+import com.devsuperior.domainModelORM.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -17,7 +22,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Recurso não encontrado !!"));
         return new ProductDTO(product);
     }
 
@@ -32,24 +38,35 @@ public class ProductService {
         Product product = new Product();
         copyDtoToEntity(productDTO, product);
         product = productRepository.save(product);
-
         return new ProductDTO(product);
     }
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product product = productRepository.getReferenceById(id);
-        copyDtoToEntity(productDTO, product);
-        productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, product);
+            productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado !!");
+        }
     }
 
-    @Transactional
-    public void delete(Long id){
-        productRepository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado !!");
+        }
+        try{
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial!!");
+        }
+
     }
 
-    private void copyDtoToEntity(ProductDTO productDTO, Product product){
+    private void copyDtoToEntity(ProductDTO productDTO, Product product) {
         product.setName(productDTO.getName());
         product.setDescription(productDTO.getDescription());
         product.setPrice(productDTO.getPrice());
